+++ src/app/app.component.html
@@ -1,8 +1,7 @@
 <div class="container">
   <nav class="navigation">
     <a mat-button routerLink="/">Team Generator</a>
-    <a mat-button routerLink="/leaderboard">Leaderboard</a>
-    <a mat-button routerLink="/extra-leaderboard">Extra Leaderboard</a>
+    <a mat-button routerLink="/extra-leaderboard">Klassement</a>
   </nav>
   <div class="content">
     <router-outlet></router-outlet>
