--- src/app/components/leaderboard/leaderboard.component.html
@@ -1,138 +1,0 @@
-<div class="leaderboard-container">
-  <mat-card>
-    <mat-card-content>
-      <mat-table [dataSource]="leaderboard" class="mat-elevation-z8 mobile-friendly-table" style="width: 100%;">
-        <!-- Position Column -->
-        <ng-container matColumnDef="position" class="position-column">
-          <mat-header-cell *matHeaderCellDef class="position-cell">
-            <span class="full-header">#</span>
-            <span class="mobile-header">#</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player; let i = index" class="position-cell">
-            {{getPlayerPosition(player)}}
-          </mat-cell>
-        </ng-container>
-
-        <!-- Games Played Column -->
-        <ng-container matColumnDef="gamesPlayed" class="games-played-column">
-          <mat-header-cell *matHeaderCellDef class="games-played-cell">
-            <span class="full-header">Wedstrijden</span>
-            <span class="mobile-header">W</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player" class="games-played-cell"> {{player.gamesPlayed}} </mat-cell>
-        </ng-container>
-
-        <!-- Total Points Column -->
-        <ng-container matColumnDef="totalPoints" class="total-points-column">
-          <mat-header-cell *matHeaderCellDef class="total-points-cell">
-            <span class="full-header">Punten</span>
-            <span class="mobile-header">P</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player" class="total-points-cell"> {{player.totalPoints}} </mat-cell>
-        </ng-container>
-
-        <!-- Games Won Column -->
-        <ng-container matColumnDef="gamesWon">
-          <mat-header-cell *matHeaderCellDef>
-            <span class="full-header">Winst</span>
-            <span class="mobile-header">W</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player"> {{player.gamesWon}} </mat-cell>
-        </ng-container>
-
-        <!-- Games Lost Column -->
-        <ng-container matColumnDef="gamesLost">
-          <mat-header-cell *matHeaderCellDef>
-            <span class="full-header">Verlies</span>
-            <span class="mobile-header">V</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player"> {{player.gamesLost}} </mat-cell>
-        </ng-container>
-
-        <!-- Games Tied Column -->
-        <ng-container matColumnDef="gamesTied">
-          <mat-header-cell *matHeaderCellDef>
-            <span class="full-header">Gelijk</span>
-            <span class="mobile-header">T</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player"> {{player.gamesTied}} </mat-cell>
-        </ng-container>
-
-        <!-- Zlatan Points Column -->
-        <ng-container matColumnDef="zlatanPoints">
-          <mat-header-cell *matHeaderCellDef>
-            <span class="full-header">Zlatan</span>
-            <span class="mobile-header">Z</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player"> {{player.zlatanPoints}} </mat-cell>
-        </ng-container>
-
-        <!-- Ventiel Column -->
-        <ng-container matColumnDef="ventielPoints">
-          <mat-header-cell *matHeaderCellDef>
-            <span class="full-header">Ventiel</span>
-            <span class="mobile-header">V</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player"> {{player.ventielPoints || 0}} </mat-cell>
-        </ng-container>
-
-        <!-- Last Five Games Column -->
-        <ng-container matColumnDef="lastFiveGames" class="last-five-games-column">
-          <mat-header-cell *matHeaderCellDef class="last-five-games-cell">
-            <span class="full-header">Laatste 5</span>
-            <span class="mobile-header">Laatste 5</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player" class="last-five-games-cell">
-            <div class="last-five-games">
-              <ng-container *ngFor="let game of getLastFiveGames(player)">
-                <mat-icon 
-                  [ngStyle]="{
-                    'color': game.result === 3 ? 'green' : game.result === 1 ? 'red' : 'gray',
-                    'margin-right': '0px'
-                  }"
-                  [matTooltip]="'Date: ' + game.date + ' - ' + (game.result === 3 ? 'Win' : game.result === 1 ? 'Loss' : 'Tie')"
-                >check_circle</mat-icon>
-              </ng-container>
-            </div>
-          </mat-cell>
-        </ng-container>
-
-        <!-- Win Ratio Column -->
-        <ng-container matColumnDef="winRatio">
-          <mat-header-cell *matHeaderCellDef>
-            <span class="full-header">Win %</span>
-            <span class="mobile-header">W%</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player"> {{player.winRatio | number:'1.1-1'}}% </mat-cell>
-        </ng-container>
-
-        <!-- Name Column with Trophy Icon for 1st Place -->
-        <ng-container matColumnDef="name" class="name-column">
-          <mat-header-cell *matHeaderCellDef class="name-cell" style="padding-right: 20px;">
-            <span class="full-header">Name</span>
-            <span class="mobile-header">Name</span>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let player; let i = index" class="name-cell" style="padding-right: 20px;">
-            {{player.name}}
-            <ng-container *ngIf="i === 0">
-              <mat-icon style="font-size: 30px; color: gold; margin-left: 5px; vertical-align: middle;">emoji_events</mat-icon>
-            </ng-container>
-            <ng-container *ngIf="i === 1">
-              <mat-icon style="font-size: 30px; color: silver; margin-left: 5px; vertical-align: middle;">emoji_events</mat-icon>
-            </ng-container>
-            <ng-container *ngIf="i === 2">
-              <mat-icon style="font-size: 30px; color: #CE8946; margin-left: 5px; vertical-align: middle;">emoji_events</mat-icon>
-            </ng-container>
-          </mat-cell>
-        </ng-container>
-
-        <mat-header-row *matHeaderRowDef="currentDisplayColumns"></mat-header-row>
-        <mat-row *matRowDef="let player; columns: currentDisplayColumns;" 
-                 (click)="openChemistryModal(player)"
-                 [class.selected-row]="player === selectedPlayer"
-                 style="cursor: pointer;">
-        </mat-row>
-      </mat-table>
-    </mat-card-content>
-  </mat-card>
-</div>
