--- src/app/components/leaderboard/leaderboard.component.scss
@@ -1,157 +1,0 @@
-// Variabelen
-$mobile-breakpoint: 767px;
-$light-hover-bg: rgba(0, 0, 0, 0.04);
-$selected-row-bg: #f5f5f5;
-
-// Algemene stijlen
-.selected-row {
-  background-color: $selected-row-bg;
-}
-
-// Header stijlen
-.full-header {
-  display: inline;
-}
-
-.mobile-header {
-  display: none;
-}
-
-// Trophy iconen in de naam kolom op desktop
-.name-cell {
-  mat-icon {
-    font-size: 30px;
-    width: 30px !important;
-    height: 30px !important;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-  }
-}
-
-// Algemene tabel verbeteringen
-mat-table {
-  overflow-x: auto;
-}
-
-mat-row {
-  &:hover {
-    background-color: $light-hover-bg;
-  }
-}
-
-// Mobiele stijlen
-@media screen and (max-width: $mobile-breakpoint) {
-  .full-header {
-    display: none;
-  }
-  
-  .mobile-header {
-    display: inline;
-  }
-
-  .leaderboard-container {
-    padding: 0;
-  }
-  
-  mat-cell, mat-header-cell {
-    padding: 8px 2px !important;
-    font-size: 13px;
-  }
-  
-  // Expliciete kolombreedte-verdeling voor mobiel
-  .mobile-friendly-table {
-    table-layout: fixed !important;
-    
-    // Override Angular Material flexbox gedrag
-    .mat-mdc-row, .mat-mdc-header-row {
-      display: flex !important;
-      
-      // Reset gelijke kolombreedte gedrag
-      .mat-mdc-cell, .mat-mdc-header-cell {
-        flex: none !important;
-      }
-    }
-    
-    // Specifieke breedtes voor alle kolommen
-    .position-cell {
-      flex: 0 0 10% !important;
-      max-width: 10% !important;
-      width: 10% !important;
-      text-align: center;
-      padding: 4px 0 !important;
-      overflow: hidden;
-      padding-left: 8px !important; // Voeg links padding toe aan de ranking kolom
-    }
-    
-    .name-cell {
-      flex: 0 0 25% !important;
-      max-width: 25% !important;
-      width: 25% !important;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-    
-    .games-played-cell, .total-points-cell {
-      flex: 0 0 10% !important;
-      max-width: 10% !important;
-      width: 10% !important;
-      text-align: center;
-      padding: 4px 0 !important;
-      overflow: hidden;
-    }
-
-    .rating-cell {
-      flex: 0 0 15% !important;
-      max-width: 15% !important;
-      width: 15% !important;
-      text-align: center;
-      padding: 4px 0 !important;
-      overflow: hidden;
-    }
-    
-    .last-five-games-cell {
-      flex: 0 0 30% !important;
-      max-width: 30% !important;
-      width: 30% !important;
-      display: flex;
-      padding: 4px 0 !important;
-      justify-content: flex-start;
-      
-      .last-five-games {
-        display: flex;
-        align-items: center;
-      }
-    }
-  }
-  
-  // Trophy en game resultaat icons aanpassingen
-  mat-icon {
-    font-size: 18px !important;
-    margin-left: 2px !important;
-  }
-  
-  .last-five-games {
-    mat-icon {
-      font-size: 16px !important;
-      margin-right: 0px !important;
-      width: 16px !important;  /* Verkleint de standaard 24px breedte naar 16px */
-      height: 16px !important; /* Verkleint de standaard 24px hoogte naar 16px */
-      display: inline-flex !important; /* Zorgt voor een betere uitlijning */
-      align-items: center !important;
-      justify-content: center !important;
-    }
-  }
-  
-  // Zorg ervoor dat de tabel alle beschikbare ruimte gebruikt
-  mat-table {
-    width: 100% !important;
-  }
-}
-
-// Extra regel om de flexbox layout direct te overschrijven
-::ng-deep .mat-mdc-table .mat-mdc-row, ::ng-deep .mat-mdc-table .mat-mdc-header-row {
-  @media screen and (max-width: $mobile-breakpoint) {
-    height: auto !important;
-  }
-}
